caption-api/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
opencv-python>=4.8.0
